requirements.txt
setuptools
requests
pyyaml
json
